requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
pg8000==1.30.3
pytz==2023.3